+++ tests/git-hooks-custom-hookspath/devenv.nix
@@ -1,0 +1,11 @@
+{
+  git-hooks.hooks = {
+    no-op = {
+      enable = true;
+      name = "No Op";
+      pass_filenames = false;
+      raw.always_run = true;
+      entry = "true";
+    };
+  };
+}